pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "parker_csaw"
version = "0.0.1"
authors = [
  { name="parker-research", email="author@example.com" },
]
description = "CSAW 2024 competition entry."
readme = "README.md"
license = "Unlicense"
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: The Unlicense (Unlicense)",
  "Operating System :: OS Independent",
]

dependencies = [
  "openai>=1.52.2,<2",
  "pyyaml>=6.0.2,<7",
  "git-python>=1.0.3,<2",
  "typed-argparse>=0.3.1,<1",
  "loguru>=0.7.2,<1",

  "ruff<1",
  "pyright",
]

[project.urls]
Homepage = "https://github.com/parker-research/"

